requirements.txt
streamlit
pandas
icrawler
pillow
imagehash